move all py files to src root
