--- src/hss/__init__.py
@@ -1,7 +1,0 @@
-
-from .part_processing import PartProcessor
-from .sketches import get_sketches_to_export
-from .bodies import get_bodies_to_export, handle_body_thickness
-from .properties import get_part_properties, get_part_properties
-
-from .tests import test
