+++ src/part_processing.py
@@ -5,9 +5,9 @@ from os import path
 import sys
 
 import config
-from nx import DxfExporter
-from .sketches import get_sketches_to_export
-from .bodies import get_bodies_to_export, handle_body_thickness
+from dxf_export import DxfExporter
+from sketches import get_sketches_to_export
+from bodies import get_bodies_to_export, handle_body_thickness
 
 import NXOpen
