+++ src/properties.py
